Crate adapter

Re-exports

pub use dummy::Dummy;

Modules

client

The Client traits that define the actions for each state.

dummy

The Dummy client for the Adapter.

ethereum

The Ethereum client for the Adapter.

prelude

Re-export of the crate::client traits and the states of the Adapter.

primitives

Primitives used by the Adapter. Including re-exported types from the primitives crate that are being used.

util

Utilities for working with the State root and the Balances

Structs

Adapter

The Adapter struct and it’s states.

Error

The error used by the crate::Adapter to wrap any custom error from the client and the kinds of errors that the crate::Adapter can return.

Ethereum

Ethereum client implementation for the crate::Adapter.

LockedState

The Locked state of the crate::Adapter. See crate::client::Locked

UnlockedState

The Unlocked state of the crate::Adapter. See crate::client::Unlocked