Module adapter::ethereum::test_util

Available on crate feature test-util only.

Structs

Erc20Token

The Mocked token API contract

Outpace

The Outpace contract

Sweeper

The Sweeper contract

Statics

GANACHE_1

GANACHE_1337

Chain of the locally running ganache-cli chain with id #1337

GANACHE_1337_WEB3

Initialized Ganache [Web3] instance using a Http transport for usage in tests. It uses the GANACHE_1337 to initialize the client.

GANACHE_INFO_1

ChainInfo of the locally running ganache-cli chain with id #1

GANACHE_INFO_1337

ChainInfo of the locally running ganache-cli chain with id #1337

IDENTITY_BYTECODE

Identity bytecode

KEYSTORES

KEYSTORE_IDENTITY

Uses local keystore.json file and it’s address for testing and working with crate::Ethereum

MOCK_TOKEN_ABI

Mocked Token ABI

MOCK_TOKEN_BYTECODE

Mocked Token bytecode in JSON

OUTPACE_BYTECODE

Outpace bytecode

SWEEPER_BYTECODE

Sweeper bytecode

Functions

deploy_identity_contract

Deploys the Identity contract for the give for_address Adds privileges by the constructor for the add_privileges_to addresses

get_test_channel