Enum adapter::flat_table::key::FlatTableKey

pub enum FlatTableKey<M> {
    Multiple {
        phantom: PhantomData<M>,
        en: Cow<'static, str>,
        de: Cow<'static, str>,
    },
    Single(Cow<'static, str>),
}

Variants

Multiple

Fields

phantom: PhantomData<M>

en: Cow<'static, str>

de: Cow<'static, str>

Single(Cow<'static, str>)

Implementations

impl FlatTableKey<Unselect>

pub const ACTIVE: FlatTableKey<Unselect> = _

pub const ADDRESS: FlatTableKey<Unselect> = _

pub const ANNOTATION: FlatTableKey<Unselect> = _

pub const BASIN_CODE: FlatTableKey<Unselect> = _

pub const COUNTY: FlatTableKey<Unselect> = _

pub const DAM_TARGETS_DEFAULT: FlatTableKey<Unselect> = _

pub const DAM_TARGETS_MAX: FlatTableKey<Unselect> = _

pub const DAM_TARGETS_STEADY: FlatTableKey<Unselect> = _

pub const DATE_OF_CHANGE: FlatTableKey<Unselect> = _

pub const EU_SURVEY_AREA: FlatTableKey<Unselect> = _

pub const EXTERNAL_IDENTIFIER: FlatTableKey<Unselect> = _

pub const FILE_REFERENCE: FlatTableKey<Unselect> = _

pub const FIRST_GRANT: FlatTableKey<Unselect> = _

pub const FLOOD_AREA: FlatTableKey<Unselect> = _

pub const FLUID_DISCHARGE: FlatTableKey<Unselect> = _

pub const GRANTING_AUTHORITY: FlatTableKey<Unselect> = _

pub const GROUNDWATER_BODY: FlatTableKey<Unselect> = _

pub const INJECTION_RATE: FlatTableKey<Unselect> = _

pub const IRRIGATION_AREA: FlatTableKey<Unselect> = _

pub const LAND_RECORD: FlatTableKey<Unselect> = _

pub const LEGAL_DEPARTMENT_ABBREVIATION: FlatTableKey<Unselect> = _

pub const LEGAL_DEPARTMENT_DESCRIPTION: FlatTableKey<Unselect> = _

pub const LEGAL_PURPOSE: FlatTableKey<Unselect> = _

pub const LEGAL_TITLE: FlatTableKey<Unselect> = _

pub const MAINTENANCE_ASSOCIATION: FlatTableKey<Unselect> = _

pub const MUNICIPAL_AREA: FlatTableKey<Unselect> = _

pub const NO: FlatTableKey<Unselect> = _

pub const PH_VALUES_MAX: FlatTableKey<Unselect> = _

pub const PH_VALUES_MIN: FlatTableKey<Unselect> = _

pub const PLOT: FlatTableKey<Unselect> = _

pub const PUMPING_RATE: FlatTableKey<Unselect> = _

pub const RAIN_SUPPLEMENT: FlatTableKey<Unselect> = _

pub const REAL: FlatTableKey<Unselect> = _

pub const REGISTERING_AUTHORITY: FlatTableKey<Unselect> = _

pub const REGULATION_CITATION: FlatTableKey<Unselect> = _

pub const RIGHTS_HOLDER: FlatTableKey<Unselect> = _

pub const RIVER_BASIN: FlatTableKey<Unselect> = _

const SORT_ORDER: [Self; 41] = _

pub const STATUS: FlatTableKey<Unselect> = _

pub const SUBJECT: FlatTableKey<Unselect> = _

pub const TOP_MAP_1_25000: FlatTableKey<Unselect> = _

pub const USAGE_LOCATION_NAME: FlatTableKey<Unselect> = _

pub const USAGE_LOCATION_NO: FlatTableKey<Unselect> = _

pub const USAGE_LOCATION_SERIAL_NO: FlatTableKey<Unselect> = _

pub const UTM_EASTING: FlatTableKey<Unselect> = _

pub const UTM_NORTHING: FlatTableKey<Unselect> = _

pub const VALID_FROM: FlatTableKey<Unselect> = _

pub const VALID_UNTIL: FlatTableKey<Unselect> = _

pub const WASTER_WATER_FLOW_VOLUME: FlatTableKey<Unselect> = _

pub const WATER_AUTHORITY: FlatTableKey<Unselect> = _

pub const WATER_BODY: FlatTableKey<Unselect> = _

pub const WATER_PROTECTION_AREA: FlatTableKey<Unselect> = _

pub const WITHDRAWAL_RATE: FlatTableKey<Unselect> = _

impl<M> FlatTableKey<M>

const fn from_str(en: &'static str, de: &'static str) -> Self

pub fn from_unselect_ref(value: &FlatTableKey<Unselect>) -> &Self

Converts a &FlatTableKey<marker::Unselect> to &FlatTableKey<M>, where M is any marker type.

Safety

This function uses std::mem::transmute to perform a zero-cost conversion of the reference. The safety of this operation is ensured because:

The memory layout of FlatTableKey<marker::Unselect> and FlatTableKey<M> is identical.

The marker types, irrespective of their differences, are encapsulated within PhantomData which does not affect the memory layout.

As such, there’s no risk of undefined behavior arising from this conversion, provided the structure of FlatTableKey remains consistent.

pub fn from_unselect(value: FlatTableKey<Unselect>) -> Self

pub fn ref_de(&self) -> &str

pub fn ref_en(&self) -> &str

impl<M> FlatTableKey<M>where FlatTableKey<M>: AsRef<str>,

pub fn sort_index(&self) -> Option<usize>