Module adapter::primitives

Primitives used by the Adapter. Including re-exported types from the primitives crate that are being used.

Structs

Address

BigNum

Chain

Ethereum Virtual Machine Chain

ChainId

The Id of the chain

ChainInfo

Configured chain with tokens.

ChainOf

Context of TokenInfo & Chain information for given Channel or Campaign.

Channel

Config

Examples:

Session

Session struct returned by the crate::Adapter when crate::client::Locked::session_from_token is called.

TokenInfo

Configured Token in a specific Chain. Precision can differ for the same token from one Chain to another.

ValidatorId

Enums

AdapterTypes

A helper type that allows you to use either of them and dereference the adapter when calling for example an application with a concrete implementation of the crate::Adapter.

Type Definitions

Deposit

The Deposit struct with BigNum values. Returned by crate::client::Locked::get_deposit